refactor(persistence): ADR-0035 — invert the audit-append write seam (enabling slice 2b)
The last crate::audit_chain coupling in verifier_store was the write path:
ChainedAuditAppender::append_within delegated to
crate::audit_chain::AuditChainLinker::append_audit_event_tx.

Rather than inject the impl up from root (which would churn every VerifierStore::new
caller), the append MECHANICS move DOWN. append_audit_event_tx writes the
persistence-owned audit_log_chain / audit_anchor_head tables using only the pure
kirra_audit_hash primitives + Ed25519 — it is storage mechanics, so it belongs in
the persistence layer. It's relocated (verbatim) into verifier_store::audit_appender
as a free fn; ChainedAuditAppender calls it locally, and root's
AuditChainLinker::append_audit_event_tx now DELEGATES down to it so its typed
wrappers (append_rss_violation / append_perception_derate), tests, and external
callers are unchanged.

No cycle (persistence -> kirra_audit_hash leaf; root -> persistence), no
store-construction churn, no new crate, no lockfile change. verifier_store now has
ZERO crate::audit_chain code references — production AND tests (the store's test
module repointed to kirra_audit_hash / the local append fn); only a prose comment
names it.

Byte-exactness is load-bearing (this is the on-disk audit-chain write). Verified
identical: power-loss audit drill (audit_chain_prefix_on_kill, 3 — reopens the WAL
and re-verifies the chain), `cargo test --lib` 760 (incl. audit verify +
break_audit_chain tamper tests), two_node_rollout EP-13 (4), fmt, guardrails,
orphan-cores.

With this, only the three small couplings (ShippedAuditRecord, is_valid_verdict_id,
KeyRegistry) stand between verifier_store and the wholesale kirra-persistence move.
